pyproject.toml
[project]
name = "kusto-mcp"
version = "0.2.0"
description = "A MCP server for Azure Data Explorer (Kusto) with zero-config headless authentication"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-kusto-data>=4.6.1",
    "azure-identity>=1.15.0",
    "mcp>=1.1.1",
]

[[project.authors]]
name = "Animesh Kundu"
email = "anik.edu@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_kusto"]

[project.scripts]
kusto-mcp = "mcp_server_kusto:main"
